+++ producer/src/app/(protected)/perfil/pessoal/page.tsx
@@ -1,0 +1,65 @@
+"use client";
+
+import { ModelPage } from "@shared/components/ModelPage";
+import PersonalInfoForm from "./components/PersonalInfoForm";
+import Button from "@shared/components/Button";
+import Link from "next/link";
+import Modal from "@shared/components/Modal";
+import { useChangeRegistrationForm } from "../components";
+
+export default function ProfilePage() {
+  const {
+    isModalOpen,
+    setIsModalOpen,
+    register,
+    confirmSubmission,
+    errors,
+    control,
+    handleSubmit,
+    handleSubmitForm,
+  } = useChangeRegistrationForm();
+
+  return (
+    <ModelPage
+      title="Seu perfil"
+      titleClassName="pt-17"
+      titleGap="gap-2.5"
+      subtitle="Após atualizar os seus dados, clique em salvar."
+      subtitleClassName="px-9 leading-5.5"
+    >
+      <form onSubmit={handleSubmit(handleSubmitForm)} className="h-full w-full flex flex-col overflow-y-hidden">
+        <div className="gap-3.5 w-full overflow-y-auto flex flex-col items-center justify-between pb-16 pt-7.5">
+          <PersonalInfoForm register={register} errors={errors} control={control} />
+        </div>
+
+        <div className="w-full flex gap-2 bg-transparent pb-5">
+          <Link className="w-full h-full" href={"/"} as={"/"}>
+            <Button className="w-full h-11 rounded-lg bg-white font-semibold text-slate-gray border-slate-gray border-2 hover:bg-gray-100">
+              Voltar
+            </Button>
+          </Link>
+          <Button
+            className="w-full h-11 rounded-lg bg-theme-default font-semibold text-white"
+            title="Salvar"
+            type="submit"
+            onClick={() => setIsModalOpen(true)}
+          >
+            Salvar
+          </Button>
+          <Modal
+            titleContentModal="Você tem certeza?"
+            contentModal="Ao clicar em confirmar seus dados de cadastro serão atualizados."
+            titleCloseModal="Cancelar"
+            titleConfirmModal="Confirmar"
+            bgOpenModal="#2F4A4D"
+            bgConfirmModal="#2F4A4D"
+            bgCloseModal="bg-gray-100"
+            modalAction={confirmSubmission}
+            isOpen={isModalOpen}
+            setIsOpen={setIsModalOpen}
+          />
+        </div>
+      </form>
+    </ModelPage>
+  );
+}
