+++ producer/src/app/_actions/farms/PATCH/update-farm.ts
@@ -2,17 +2,12 @@
 
 import ApiService from "@shared/service";
 
-export interface UpdateFarmRequest {
-  name: string;
-  description?: string;
-}
-
-export async function updateFarm({ name, description }: UpdateFarmRequest) {
+export async function updateFarm(data: FormData) {
   const response = ApiService.PATCH({
     url: "/farms/own",
-    data: {
-      name,
-      description,
+    data,
+    headers: {
+      "Content-Type": "multipart/form-data",
     },
   });
